Refactor:Extract first class collection
